refactor: split docs markdown linking helpers
