+++ crates/ox_content_docs/src/markdown/regex_cache.rs
@@ -1,0 +1,16 @@
+use std::sync::OnceLock;
+
+use regex::Regex;
+
+pub(super) type RegexCache = OnceLock<Option<Regex>>;
+
+pub(super) fn cached_regex(
+    cache: &'static RegexCache,
+    pattern: &'static str,
+) -> Option<&'static Regex> {
+    // Regex construction is expensive and these helpers run throughout doc
+    // generation. Cache both success and failure in `OnceLock<Option<_>>` so a
+    // bad pattern degrades to the fallback path without recompiling on every
+    // call.
+    cache.get_or_init(|| Regex::new(pattern).ok()).as_ref()
+}
